refactor: fetch metrics for search on services
