+++ src/rubrix/server/services/tasks/text2text/service.py
@@ -17,7 +17,7 @@ from typing import Iterable, List, Optional, Type
 
 from fastapi import Depends
 
-from rubrix.server.services.metrics import ServiceBaseMetric
+from rubrix.server.commons.config import TasksFactory
 from rubrix.server.services.metrics.models import ServiceBaseTaskMetrics
 from rubrix.server.services.search.model import (
     ServiceSearchResults,
@@ -84,29 +84,19 @@ class Text2TextService:
         record_from: int = 0,
         size: int = 100,
         exclude_metrics: bool = True,
-        metrics: Optional[List[ServiceBaseMetric]] = None,
     ) -> ServiceSearchResults:
-        """
-        Run a search in a dataset
 
-        Parameters
-        ----------
-        dataset:
-            The records dataset
-        query:
-            The search parameters
-        sort_by:
-            The sort by list
-        record_from:
-            The record from return results
-        size:
-            The max number of records to return
-
-        Returns
-        -------
-            The matched records with aggregation info for specified task_meta.py
-
-        """
+        metrics = TasksFactory.find_task_metrics(
+            dataset.task,
+            metric_ids={
+                "words_cloud",
+                "predicted_by",
+                "annotated_by",
+                "status_distribution",
+                "metadata",
+                "score",
+            },
+        )
 
         results = self.__search__.search(
             dataset,
