+++ src/rubrix/server/services/tasks/text_classification/service.py
@@ -17,8 +17,8 @@ from typing import Iterable, List, Optional, Type
 
 from fastapi import Depends
 
+from rubrix.server.commons.config import TasksFactory
 from rubrix.server.errors.base_errors import MissingDatasetRecordsError
-from rubrix.server.services.metrics import ServiceBaseMetric
 from rubrix.server.services.metrics.models import ServiceBaseTaskMetrics
 from rubrix.server.services.search.model import (
     ServiceSearchResults,
@@ -93,7 +93,6 @@ class TextClassificationService:
         record_from: int = 0,
         size: int = 100,
         exclude_metrics: bool = True,
-        metrics: Optional[List[ServiceBaseMetric]] = None,
     ) -> ServiceSearchResults:
         """
         Run a search in a dataset
@@ -115,6 +114,21 @@ class TextClassificationService:
             The matched records with aggregation info for specified task_meta.py
 
         """
+
+        metrics = TasksFactory.find_task_metrics(
+            dataset.task,
+            metric_ids={
+                "words_cloud",
+                "predicted_by",
+                "predicted_as",
+                "annotated_by",
+                "annotated_as",
+                "error_distribution",
+                "status_distribution",
+                "metadata",
+                "score",
+            },
+        )
 
         results = self.__search__.search(
             dataset,
